+++ cypress/locators/flash/locators.js
@@ -1,3 +1,2 @@
 // component preview locators
 export const FLASH_PREVIEW = '[data-component="flash"]';
-export const MESSAGE_PREVIEW = '[data-element="message"]';