+++ src/__experimental__/components/radio-button/radio-button-group.component.js
@@ -6,6 +6,7 @@ import Fieldset from '../../../__internal__/fieldset';
 import RadioButtonGroupStyle from './radio-button-group.style';
 import RadioButtonMapper from './radio-button-mapper.component';
 import Logger from '../../../utils/logger/logger';
+import useIsAboveBreakpoint from '../../../hooks/__internal__/useIsAboveBreakpoint';
 
 let deprecatedWarnTriggered = false;
 
@@ -16,9 +17,40 @@ const RadioButtonGroup = (props) => {
     Logger.deprecate('`styleOverride` that is used in the `RadioButtonGroup` component is deprecated and will soon be removed.');
   }
   const {
-    children, name, legend, error, warning, info, onBlur,
-    onChange, value, inline, legendInline, styleOverride
+    children,
+    name,
+    legend,
+    error,
+    warning,
+    info,
+    onBlur,
+    onChange,
+    value,
+    inline = false,
+    legendInline = false,
+    legendWidth,
+    legendAlign,
+    legendSpacing,
+    ml,
+    mb,
+    labelSpacing = 1,
+    adaptiveLegendBreakpoint,
+    adaptiveSpacingBreakpoint,
+    styleOverride = {}
   } = props;
+
+  const isAboveLegendBreakpoint = useIsAboveBreakpoint(adaptiveLegendBreakpoint);
+  const isAboveSpacingBreakpoint = useIsAboveBreakpoint(adaptiveSpacingBreakpoint);
+
+  let inlineLegend = legendInline;
+  if (adaptiveLegendBreakpoint) {
+    inlineLegend = isAboveLegendBreakpoint;
+  }
+
+  let marginLeft = ml;
+  if (adaptiveSpacingBreakpoint && !isAboveSpacingBreakpoint) {
+    marginLeft = undefined;
+  }
 
   return (
     <Fieldset
@@ -27,7 +59,12 @@ const RadioButtonGroup = (props) => {
       error={ error }
       warning={ warning }
       info={ info }
-      inline={ legendInline }
+      inline={ inlineLegend }
+      legendWidth={ legendWidth }
+      legendAlign={ legendAlign }
+      legendSpacing={ legendSpacing }
+      ml={ marginLeft }
+      mb={ mb }
       styleOverride={ styleOverride }
       { ...tagComponent('radiogroup', props) }
     >
@@ -45,6 +82,7 @@ const RadioButtonGroup = (props) => {
         >
           {React.Children.map(children, child => React.cloneElement(child, {
             inline,
+            labelSpacing,
             error: !!error,
             warning: !!warning,
             info: !!info,
@@ -85,18 +123,28 @@ RadioButtonGroup.propTypes = {
   inline: PropTypes.bool,
   /** When true, legend is placed in line with the radiobuttons */
   legendInline: PropTypes.bool,
+  /** Percentage width of legend (only when legend is inline)  */
+  legendWidth: PropTypes.number,
+  /** Text alignment of legend when inline */
+  legendAlign: PropTypes.oneOf(['left', 'right']),
+  /** Spacing between legend and field for inline legend, number multiplied by base spacing unit (8) */
+  legendSpacing: PropTypes.oneOf([1, 2]),
+  /** Margin left, any valid CSS value */
+  ml: PropTypes.string,
+  /** Margin bottom, given number will be multiplied by base spacing unit (8) */
+  mb: PropTypes.oneOf([0, 1, 2, 3, 4, 5, 6, 7]),
+  /** Spacing between labels and radio buttons, given number will be multiplied by base spacing unit (8) */
+  labelSpacing: PropTypes.oneOf([1, 2]),
+  /** Breakpoint for adaptive legend (inline labels change to top aligned). Enables the adaptive behaviour when set */
+  adaptiveLegendBreakpoint: PropTypes.number,
+  /** Breakpoint for adaptive spacing (left margin changes to 0). Enables the adaptive behaviour when set */
+  adaptiveSpacingBreakpoint: PropTypes.number,
   /** Allows to override existing component styles */
   styleOverride: PropTypes.shape({
     root: PropTypes.oneOfType([PropTypes.func, PropTypes.object]),
     content: PropTypes.oneOfType([PropTypes.func, PropTypes.object]),
     legend: PropTypes.oneOfType([PropTypes.func, PropTypes.object])
   })
-};
-
-RadioButtonGroup.defaultProps = {
-  inline: false,
-  legendInline: false,
-  styleOverride: {}
 };
 
 export default RadioButtonGroup;