+++ src/__experimental__/components/switch/switch.style.js
@@ -12,7 +12,7 @@ import ClassicSwitchStyles from './switch-classic.style';
 
 const StyledSwitch = styled.div`
   ${({
-    fieldHelpInline, labelInline, labelWidth, reverse, size, theme
+    fieldHelpInline, labelInline, labelSpacing, reverse, size, theme
   }) => css`
     ${FieldLineStyle} {
       display: flex;
@@ -28,8 +28,6 @@ const StyledSwitch = styled.div`
       margin-left: 0;
     }
 
-    margin-bottom: 24px;
-
     ${HiddenCheckableInputStyle}:not([disabled]) {
       &:focus + ${StyledSwitchSlider},
       &:hover + ${StyledSwitchSlider} {
@@ -42,14 +40,13 @@ const StyledSwitch = styled.div`
     }
 
     ${StyledLabelContainer} {
+      ${!labelInline && css`padding: 0;`};
+      margin-bottom: 8px;
+
       ${StyledValidationIcon} {
         position: relative;
         display: inline-block;
       }
-
-      ${labelWidth && css`
-        margin-right: ${100 - labelWidth}%;
-      `}
     }
 
     ${fieldHelpInline && css`
@@ -83,9 +80,6 @@ const StyledSwitch = styled.div`
 
       ${StyledLabelContainer} {
         margin-bottom: 0;
-        padding-right: 0;
-        margin-right: 32px;
-        width: auto;
       }
 
       ${FieldHelpStyle} {
@@ -99,13 +93,10 @@ const StyledSwitch = styled.div`
           margin-top: 0;
         }
 
-        ${StyledLabelContainer} {
-          margin-left: 10px;
-        }
-
         ${!fieldHelpInline && `
           ${FieldHelpStyle} {
-            margin-left: 70px;
+            margin-left: 60px;
+            padding-left: ${labelSpacing * theme.spacing}px;
           }
         `}
       `}
@@ -136,8 +127,23 @@ const StyledSwitch = styled.div`
 
       ${labelInline && !fieldHelpInline && reverse && css`
         ${FieldHelpStyle} {
-          margin-left: 88px;
+          padding: 10px 0;
         }
+      `}
+
+      ${labelInline && css`
+        ${StyledLabelContainer} {
+          margin-top: 1px;
+          padding-top: 10px;
+          padding-bottom: 10px;
+        }
+
+        ${!fieldHelpInline && reverse && `
+          ${FieldHelpStyle} {
+            margin-left: 78px;
+            padding-left: ${labelSpacing * theme.spacing}px;
+          }
+        `}
       `}
     `}