+++ src/components/definition-list/dt.component.js
@@ -1,0 +1,28 @@
+import React from 'react';
+import propTypes from '@styled-system/prop-types';
+import PropTypes from 'prop-types';
+import { StyledDt } from './definition-list.style';
+
+const Dt = ({
+  pr = 3,
+  children,
+  mb = 2
+}) => {
+  return (
+    <StyledDt
+      data-element='dt'
+      mb={ mb }
+      pr={ pr }
+    >
+      {children}
+    </StyledDt>
+  );
+};
+
+Dt.propTypes = {
+  ...propTypes.space,
+  /** prop to render string for `<Dt />` */
+  children: PropTypes.string.isRequired
+};
+
+export default Dt;